Support for Ceph Export and Bugfixes
Support added for export to ceph

Bugfixes for handling DNG zip files

hp tool support added for 3d model.sc.zip

Bugfixes for hp tool GUI on macOS

Bugfixes for issue with AudioCompare 'controller' keyword

Updated Operations dictionary

+++ scripts/python/jtprefs.py
@@ -1,4 +1,5 @@
 import os
+import tempfile
 import tkFileDialog
 from Tkinter import *
 import tkMessageBox
@@ -8,23 +9,35 @@ import subprocess
 import maskgen.maskgen_loader
 from maskgen.software_loader import getFileName
 from tkinter import ttk
+from hp.hp_data import orgs
 
 key = os.path.join(os.path.expanduser("~"), "medifor_ingest.gpg")
 hp_settings = os.path.join(os.path.expanduser("~"), ".hpsettings")
 
 
 class Window(Frame):
-    def __init__(self, parent, errors, gpg):
+    def __init__(self, parent, errors):
         Frame.__init__(self, parent)
         self.parent = parent
         self.parent.title("Settings")
-        self.valid_usernames = []
-        self.gpg = gpg
+        self.loader = maskgen.maskgen_loader.MaskGenLoader()
         self.setup_window()
         maskgen.maskgen_loader.imageLoaded = False
-        self.loader = maskgen.maskgen_loader.MaskGenLoader()
         if errors:
             tkMessageBox.showerror("Error", "\n".join(errors))
+
+        self.info = {"username": ["Username Field", "Enter your project codename."],
+                     "organization": ["Organization Field", "Enter the organization you are affiliated with."],
+                     "apiurl": ["API URL Field", "Enter the API URL for the browser."],
+                     "busername": ["Browser Username Field", "Enter your browser username."],
+                     "bpassword": ["Browser Password Field", "Enter your browser password."],
+                     "hporganization": ["HP Organization Field", "Enter your organization abbreviation for the HP Tool."],
+                     "uploadfolder": ["Folder Field", "Enter the location you would like to upload the tar files to."
+                                                      "\n\"s3://\" is not necessary."],
+                     "s3-endpoint": ["AWS endpoint URL field", "Enter your endpoint url if you have one."],
+                     "s3-profile": ["AWS profile name field", "Enter your aws profile name if you have multiple config profiles."],
+                     "s3-region": ["AWS region field", "Enter your aws region if you have one."],
+                     "help": ["Help", "For additional help contact MediFor_Manipulators@partech.com."]}
 
     def setup_window(self):
         r = 0
@@ -56,21 +69,22 @@ class Window(Frame):
         # API URL
         self.apiurl_label = Button(text="API URL*", command=lambda: self.get_info("apiurl"))
         self.apiurl_label.grid(row=r, column=0, padx=10)
-        self.apiurl_field = Entry()
+        self.apiurl_field = Entry(self.parent)
+        self.apiurl_field.insert(0, self.loader.get_key('apiurl', ''))
         self.apiurl_field.grid(row=r, column=1, padx=10)
         r += 1
 
         # Browser Username
         self.busername_label = Button(text="Browser Username*", command=lambda: self.get_info("busername"))
         self.busername_label.grid(row=r, column=0, padx=10)
-        self.busername_field = Entry()
+        self.busername_field = Entry(self.parent)
         self.busername_field.grid(row=r, column=1, padx=10)
         r += 1
 
         # Browser Password
         self.bpassword_label = Button(text="Browser Password*", command=lambda: self.get_info("bpassword"))
         self.bpassword_label.grid(row=r, column=0, padx=10)
-        self.bpassword_field = Entry(show="*")
+        self.bpassword_field = Entry(self.parent, show="*")
         self.bpassword_field.grid(row=r, column=1, padx=10)
         r += 1
 
@@ -78,6 +92,7 @@ class Window(Frame):
         self.username_label = Button(text="Username*", command=lambda: self.get_info("username"))
         self.username_label.grid(row=r, column=0, padx=10)
         self.username_field = ttk.Combobox(values=self.valid_usernames)
+        self.username_field.insert(0, self.loader.get_key('username', ''))
         self.username_field.grid(row=r, column=1, padx=10)
         r += 1
 
@@ -89,14 +104,16 @@ class Window(Frame):
         # Organization
         self.organization_label = Button(text="Organization*", command=lambda: self.get_info("organization"))
         self.organization_label.grid(row=r, column=0, padx=10)
-        self.organization_field = Entry()
+        self.organization_field = Entry(self.parent)
+        self.organization_field.insert(0, self.loader.get_key('organization', ''))
         self.organization_field.grid(row=r, column=1, padx=10)
         r += 1
 
         # Journal Upload Folder
         self.jt_uploadfolder_label = Button(text="Journal Upload Folder", command=lambda: self.get_info("uploadfolder"))
         self.jt_uploadfolder_label.grid(row=r, column=0, padx=10)
-        self.jt_uploadfolder_field = Entry()
+        self.jt_uploadfolder_field = Entry(self.parent)
+        self.jt_uploadfolder_field.insert(0, self.loader.get_key('s3info', ''))
         self.jt_uploadfolder_field.grid(row=r, column=1, padx=10)
         r += 1
 
@@ -105,10 +122,20 @@ class Window(Frame):
         jt_setup.grid(row=r, columnspan=2, pady=5)
         r += 1
 
+        # HP Organization
+        hporg_button = Button(text="HP Organization*", command=lambda: self.get_info("hporganization"))
+        hporg_button.grid(row=r, column=0, padx=10)
+        self.hporganization = StringVar()
+        self.hporganization.set(self.loader.get_key('hporganization', ''))
+        hporg_optionmenu = OptionMenu(None, self.hporganization, *orgs.keys())
+        hporg_optionmenu.grid(row=r, column=1, padx=10)
+        r += 1
+
         # High Provenance Upload Folder
         self.hpupload_button = Button(text="HP Upload Folder", command=lambda: self.get_info("uploadfolder"))
         self.hpupload_button.grid(row=r, column=0, padx=10)
         self.hpupload_field = Entry(self.parent)
+        self.hpupload_field.insert(0, self.loader.get_key('aws-hp', ''))
         self.hpupload_field.grid(row=r, column=1, padx=10)
         r += 1
 
@@ -116,7 +143,32 @@ class Window(Frame):
         self.prnuupload_button = Button(text="PRNU Upload Folder", command=lambda: self.get_info("uploadfolder"))
         self.prnuupload_button.grid(row=r, column=0, padx=10)
         self.prnuupload_field = Entry(self.parent)
+        self.prnuupload_field.insert(0, self.loader.get_key('aws-prnu', ''))
         self.prnuupload_field.grid(row=r, column=1, padx=10)
+        r += 1
+
+        # AWS Profile
+        self.profile_button = Button(text="AWS Profile Name", command=lambda: self.get_info("s3-profile"))
+        self.profile_button.grid(row=r, column=0, padx=10)
+        self.profile_field = Entry(self.parent)
+        self.profile_field.insert(0, self.loader.get_key('s3-profile', 'default'))
+        self.profile_field.grid(row=r, column=1, padx=10)
+        r+=1
+
+        # AWS Endpoint
+        self.endpoint_button = Button(text="AWS Endpoint URL", command=lambda: self.get_info("s3-endpoint"))
+        self.endpoint_button.grid(row=r, column=0, padx=10)
+        self.endpoint_field = Entry(self.parent)
+        self.endpoint_field.insert(0, self.loader.get_key('s3-endpoint', ''))
+        self.endpoint_field.grid(row=r, column=1, padx=10)
+        r += 1
+
+        # AWS Region
+        self.region_button = Button(text="AWS Region", command=lambda: self.get_info("s3-region"))
+        self.region_button.grid(row=r, column=0, padx=10)
+        self.region_field = Entry(self.parent)
+        self.region_field.insert(0, self.loader.get_key('s3-region', 'us-east-1'))
+        self.region_field.grid(row=r, column=1, padx=10)
         r += 1
 
         # Submit Button
@@ -128,21 +180,7 @@ class Window(Frame):
         help.grid(row=r, column=1, padx=10, pady=5)
 
     def get_info(self, item):
-        if item == "username":
-            tkMessageBox.showinfo("Username Field", "Enter your project codename.")
-        elif item == "organization":
-            tkMessageBox.showinfo("Organization Field", "Enter the organization you are affiliated with.")
-        elif item == "apiurl":
-            tkMessageBox.showinfo("API URL Field", "Enter the API URL for the browser.")
-        elif item == "busername":
-            tkMessageBox.showinfo("Browser Username Field", "Enter your browser username.")
-        elif item == "bpassword":
-            tkMessageBox.showinfo("Browser Password Field", "Enter your browser password.")
-        elif item == "uploadfolder":
-            tkMessageBox.showinfo("Folder Field", "Enter the location you would like to upload the tar files to."
-                                                  "\n\"s3://\" is not necessary.")
-        elif item == "help":
-            tkMessageBox.showinfo("Help", "For additional help contact MediFor_Manipulators@partech.com.")
+        tkMessageBox.showinfo(*self.info[item])
 
     def submit_data(self):
         self.username = self.username_field.get()
@@ -153,9 +191,14 @@ class Window(Frame):
         self.jt_uploadfolder = self.jt_uploadfolder_field.get()
         self.hpupload_folder = self.hpupload_field.get()
         self.prnuupload_folder = self.prnuupload_field.get()
+        self.s3_profile = self.profile_field.get()
+        self.s3_endpoint = self.endpoint_field.get()
+        self.s3_region = self.region_field.get()
         self.eemail = self.get_recipient()
+        self.full_org = self.hporganization.get() + " (" + orgs[self.hporganization.get()] + ")"
 
-        if not all([self.username, self.organization, self.apiurl, self.busername, self.bpassword]):
+        if not all([self.username, self.organization, self.apiurl, self.busername, self.bpassword,
+                    self.hporganization.get()]):
             tkMessageBox.showerror("Missing Fields", "One or more fields are missing required information.")
             return
 
@@ -176,17 +219,17 @@ class Window(Frame):
         if not os.path.isfile(key):
             return None
 
-        if self.gpg:
-            try:
-                gpg_result = subprocess.Popen([self.gpg, "--with-colons", key], stdout=subprocess.PIPE,
-                                              stderr=subprocess.PIPE).communicate()
-            except WindowsError:
-                tkMessageBox.showerror("Error", "There has been an error retrieving the encryption key.")
+        try:
+            gpg_result = subprocess.Popen(["gpg", "--with-colons", key], stdout=subprocess.PIPE,
+                                          stderr=subprocess.PIPE).communicate()
+        except Exception:
+            tkMessageBox.showerror("Error", "There has been an error retrieving the encryption key.")
+            return
 
-            for line in gpg_result[0].split("\n"):
-                if line.startswith("uid"):
-                    email = line.split("<")[1].split(">")[0]
-                    return email
+        for line in gpg_result[0].split("\n"):
+            if line.startswith("uid"):
+                email = line.split("<")[1].split(">")[0]
+                return email
         return None
 
     def get_token(self):
@@ -209,7 +252,10 @@ class Window(Frame):
     def create_json(self):
         data = {"username": self.username, "apitoken": self.apitoken, "organization": self.organization,
                 "s3info": self.jt_uploadfolder, "apiurl": self.apiurl, "archive_recipient": self.eemail, "aws-hp":
-                    self.hpupload_folder, "aws-prnu": self.prnuupload_folder}
+                self.hpupload_folder, "aws-prnu": self.prnuupload_folder, "autosave": "600", "fullorgname":
+                self.full_org, "hp-organization": orgs[self.hporganization.get()], "git.branch": branch,
+                "s3-endpoint": self.s3_endpoint, "s3-profile": self.s3_profile, "s3-region": self.s3_region}
+
         self.loader.saveall(data.items())
 
     def get_maskgen(self):
@@ -230,6 +276,10 @@ def update_user_name():
     import json
     from maskgen.software_loader import getFileName
     property_file = getFileName('project_properties.json')
+
+    if property_file is None:
+        return
+
     with open(property_file, 'r') as f:
         props = json.load(f)
         for prop in props['properties']:
@@ -239,31 +289,41 @@ def update_user_name():
         json.dump(props, f, indent=2, encoding='utf-8')
 
 
-def setup(gpg):
+def setup():
     errors = []
 
     if os.path.isfile(key):
         try:
-            if gpg is not None:
-                key_installed = subprocess.Popen([gpg, "--list-keys", key], stdout=subprocess.PIPE,
-                                                 stderr=subprocess.PIPE).communicate()
-                if not key_installed[0] or len(key_installed[0]) < 3:
-                    subprocess.Popen([gpg, "--import", key])
-                else:
-                    os.remove(key)
+            key_installed = subprocess.Popen(["gpg", "--list-keys", key], stdout=subprocess.PIPE,
+                                             stderr=subprocess.PIPE, shell=True)
+            key_installed.communicate()
+            if key_installed.returncode == 2:
+                subprocess.Popen(["gpg", "--import", key])
+            else:
+                os.remove(key)
         except WindowsError as e:
-            errors.append("Error encountered while installing encryption key " + str(e))
+            errors.append("Error encountered while installing encryption key: " + str(e))
 
-    if os.path.isfile(os.path.join(os.path.expanduser("~"), "Desktop", "JT.cmd")):
-        os.remove(os.path.join(os.path.expanduser("~"), "Desktop", "JT.cmd"))
-    with open(os.path.join(os.path.expanduser("~"), "Desktop", "JT.cmd"), "a+") as startjt:
-        startjt.writelines(["title Journaling Tool\n", "cd {0}\n".format(os.path.expanduser("~")), "cls\n", "jtui"])
-    if os.path.isfile(os.path.join(os.path.expanduser("~"), "Desktop", "HP_Tool.cmd")):
-        os.remove(os.path.join(os.path.expanduser("~"), "Desktop", "HP_Tool.cmd"))
-    with open(os.path.join(os.path.expanduser("~"), "Desktop", "HP_Tool.cmd"), "a+") as starthp:
-        starthp.writelines(["title HP Tool\n", "cd {0}\n".format(os.path.expanduser("~")), "cls\n", "hpgui"])
+    # Set autosave to 600s by default
+    maskgen.maskgen_loader.imageLoaded = False
+    settings = maskgen.maskgen_loader.MaskGenLoader()
+    autosave = settings.get_key("autosave")
+    if autosave is None:
+        settings.save("autosave", "600")
+
+    if sys.platform.startswith("win"):
+        # Will only run if .maskgen2 doesn't exist, so delete the old commands
+        if os.path.isfile(os.path.join(os.path.expanduser("~"), "Desktop", "JT.cmd")):
+            os.remove(os.path.join(os.path.expanduser("~"), "Desktop", "JT.cmd"))
+        with open(os.path.join(os.path.expanduser("~"), "Desktop", "JT.cmd"), "a+") as startjt:
+            startjt.writelines(["title Journaling Tool\n", "cd {0}\n".format(os.path.expanduser("~")), "cls\n",
+                                "jtui"])
+        if os.path.isfile(os.path.join(os.path.expanduser("~"), "Desktop", "HP_Tool.cmd")):
+            os.remove(os.path.join(os.path.expanduser("~"), "Desktop", "HP_Tool.cmd"))
+        with open(os.path.join(os.path.expanduser("~"), "Desktop", "HP_Tool.cmd"), "a+") as starthp:
+            starthp.writelines(["title HP Tool\n", "cd {0}\n".format(os.path.expanduser("~")), "cls\n", "hpgui"])
+
     update_user_name()
-
     return errors
 
 
@@ -298,18 +358,30 @@ def combine_settings():
     os.remove(hp_settings)
 
 
-if __name__ == "__main__":
-    gpg_path = "C:\\Program Files (x86)\\GnuPG\\bin\\gpg.exe"
-
-    errs = setup(gpg_path)
+def main():
+    root = Tk()
 
     if os.path.isfile(hp_settings):
         combine_settings()
 
     if os.path.isfile(os.path.join(os.path.expanduser("~"), ".maskgen2")):
-        exit(0)
+        # Get a maskgen loader to check if fields are defined
+        maskgen.maskgen_loader.imageLoaded = False
+        loader = maskgen.maskgen_loader.MaskGenLoader()
+        if "apitoken" in loader:
+            exit(0)
+        if "git.branch" in loader:
+            global branch
+            branch = loader.get_key("git.branch")
+        maskgen.maskgen_loader.imageLoaded = False
 
-    root = Tk()
-    Window(root, errs, gpg_path)
+    errs = setup()
+
+    Window(root, errs)
     root.wm_resizable(width=FALSE, height=FALSE)
     root.mainloop()
+
+
+if __name__ == "__main__":
+    branch = "master"
+    main()
